+++ server/api/index.js
@@ -3,7 +3,7 @@ module.exports = router
 
 router.use('/users', require('./users'))
 router.use('/products', require('./products'))
-router.use('/cart', require('./cart'))
+router.use('/order', require('./order'))
 router.use('/reviews', require('./reviews'))
 
 router.use((req, res, next) => {
