+++ server/db/models/index.js
@@ -1,7 +1,7 @@
 const User = require('./user')
 const Review = require('./review')
 const Product = require('./product')
-const Order = require('./cart')
+const Order = require('./order')
 
 /**
  * If we had any associations to make, this would be a great place to put them!
